+++ bin/tasks/dependencies.sh
@@ -4,3 +4,7 @@ cd ${PROJECT_THEME_PATH}
 rm -f package-lock.json
 npm update
 npm install
+
+# Ensure husky hooks are configured (DEV-742)
+echo "${YELLOW}Configuring pre-commit hooks...${TXTRESET}"
+node node_modules/@digitoimistodude/code-quality-checks/bin/setup.js